Version beta12
Added SVK support.

Replaced SVN-specific command SVNInfo with VCSInfo, which is defined for SVK
and SVN.  This is mapped to '<leader>ci' by default; as a consequence, the
default mapping for the CVS-specific CVSEditors command was changed to
'<leader>cE'.

Made VCSAnnotate accept parameters to pass to the underlying VCS.

Made error messages for operations on non-versioned files more consistent.

Added check to disable individual VCS extension plugins.