Version beta22
Added VCSCommandGitDescribeArgList option to control allowed modes of 'git describe' used in GetBufferInfo for git.  This is a comma-separated list of arguments to try against git-describe.  This is an attempt to have prioritized fallbacks of descriptions, and runs in order until it finds a valid description.  This value defaults to ',tags,all,always', and so first searches annotated tags, then tags, then all references, then a commit description.